+++ pkg/engine/mutate/patch/patchesUtils.go
@@ -139,7 +139,8 @@ func ignorePatch(path string) bool {
 			!strings.Contains(path, "/metadata/namespace") &&
 			!strings.Contains(path, "/metadata/annotations") &&
 			!strings.Contains(path, "/metadata/labels") &&
-			!strings.Contains(path, "/metadata/ownerReferences") {
+			!strings.Contains(path, "/metadata/ownerReferences") &&
+			!strings.Contains(path, "/metadata/finalizers") {
 			return true
 		}
 	}